Update cp_main.yml
Added sheduled run to run at eveyday 00:00.
But as I don't know how to generate tags and release inside the actions, the artifact will be uploaded. 

Means if a new version found during sheduled run the artifact of patched app only be in uploaded artifacts in github actions. Won't release.